code/streamlit/requirements.txt
streamlit
praw
python-dotenv
sumy
transformers
torch
openai